pyproject.toml
[tool.poetry]
name = "golubator-back"
version = "0.1.0"
description = ""
authors = ["2happy <biv888iii888@yandex.ru>"]
license = "No"
readme = "README.md"

package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.8"
pyjwt = "^2.10.1"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"
pymongo = "^4.11.1"
motor = "^3.7.0"

[tool.poetry.group.dev.dependencies]
poetry-plugin-export = "^1.9.0"
pre-commit = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
